% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{A study of indoor air pollution and respiratory disease}
\format{A data frame with columns:
\describe{
 \item{id}{the subject identification number.}
 \item{visit}{the visit number which can be 0, 1, or 2.}
 \item{room}{the room in which the monitor was placed.}
 \item{value}{ the level of pollution in micrograms per cubic meter.}
 \item{timepoint}{the time point of the monitor value for a given visit/room.}
}}
\source{
Homework2 Statistical Computing(140.776.01)
}
\usage{
data
}
\description{
A study of indoor air pollution and respiratory disease
}
\examples{
\dontrun{
 data
}
}
\keyword{datasets}
