% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{make_LD}
\alias{make_LD}
\title{Construct a LongitudinalData object}
\usage{
make_LD(x)
}
\arguments{
\item{x}{A dataframe that users want to convert it into a LongitudinalData object}
}
\value{
This function returns a LongitudinalData object
}
\description{
This is a function that take a dataframe as a parameter, and then use the
default setting to return a LongitudinalData object.
}
