% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\docType{methods}
\name{print,LongitudinalData-method}
\alias{print,LongitudinalData-method}
\title{Print the LongitudinalData object}
\usage{
\S4method{print}{LongitudinalData}(x)
}
\arguments{
\item{x}{a LongitudinalData object that the user wants to print}
}
\value{
This function has no return
}
\description{
If the LongitudinalData object is the result of one query, such as a certain subject, a certain visit or a certain room,
then the summary result will be a spread table of the values of the variable not included. For instance, if the object
is the result of querying a subject, then the summary result will be the spread table of visit and room. If the Longitud
inalData object is not the result of a single query, then the result of the summary will be the quantile of the values,
which includes 'Min.', '1st Qu.', "Median", "Mean", "3rd Qu.", 'Max.'.
}
