% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{print}
\alias{print}
\title{Generic function for print}
\usage{
print(x, ...)
}
\arguments{
\item{x}{It's the LongitudinalData that will be used}

\item{...}{the content that the user wants to print}
}
\description{
This is the generic function for print
}
