% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\docType{methods}
\name{subject,LongitudinalData-method}
\alias{subject,LongitudinalData-method}
\title{Extract a subject from the whole LongitudinalData object}
\usage{
\S4method{subject}{LongitudinalData}(x, query)
}
\arguments{
\item{x}{A LongitudinalData object that users want to extract specific samples from}

\item{query}{The specific subject that the user wants to know}
}
\value{
This function returns a LongitudinalData object for the specific subject
}
\description{
This is a function that take a LongitudinalData object as a parameter, and then use the
query parameter to find the correspongding subject
}
