% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{subject}
\alias{subject}
\title{Generic Function for subject}
\usage{
subject(x, ...)
}
\arguments{
\item{x}{It's the Longitudinal object that user wants to investigate}

\item{...}{It's the query id that will be used}
}
\description{
This is the generic function for subject method
}
