% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\docType{methods}
\name{summary,LongitudinalData-method}
\alias{summary,LongitudinalData-method}
\title{Obtain the summary result for a LongitudinalData object}
\usage{
\S4method{summary}{LongitudinalData}(x)
}
\arguments{
\item{x}{a LongitudinalData object that the user wants to know its summary information}
}
\value{
This function returns a summary dataframe
}
\description{
If the LongitudinalData object is the result of one query, such as a certain subject, a certain visit or a certain room,
then the summary result will be a spread table of the values of the variable not included. For instance, if the object
is the result of querying a subject, then the summary result will be the spread table of visit and room. If the Longitud
inalData object is not the result of a single query, then the result of the summary will be the quantile of the values,
which includes 'Min.', '1st Qu.', "Median", "Mean", "3rd Qu.", 'Max.'.
}
