% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{summary}
\alias{summary}
\title{Generic Function for summary}
\usage{
summary(x)
}
\arguments{
\item{x}{It's the LongitudinalData that will be used}
}
\description{
This is the generic function for summary method
}
